+++ src/createNewList.tsx
@@ -1,0 +1,5 @@
+import { NewEffortForm } from "./components/NewEffortForm";
+
+export default function Command() {
+  return <NewEffortForm />;
+}
