+++ src/globals.ts
@@ -1,0 +1,15 @@
+export const intensities = [
+    {
+      name: "On",
+      icon: "🔥",
+    },
+    {
+      name: "Ongoing",
+      icon: "♻️",
+    },
+    {
+      name: "Simmering",
+      icon: "〰︎",
+    },
+  ];
+
